Add mainmenuPage, grop page wigget, change old main.dart to loginpage.dart, move func main from loginpage to new mani
